+++ 1.10.2/src/main/java/com/stateshifterlabs/achievementbooks/ClientProxy.java
@@ -24,4 +24,9 @@ public class ClientProxy extends CommonProxy {
 		return Minecraft.getMinecraft().mcDataDir.getAbsolutePath();
 	}
 
+	@Override
+	public String getPlayerName() {
+		return Minecraft.getMinecraft().thePlayer.getName();
+	}
+
 }