+++ 1.10.2/src/main/java/com/stateshifterlabs/achievementbooks/CommonProxy.java
@@ -13,4 +13,8 @@ public class CommonProxy {
 
 	public void refreshResources() { }
 
+	public String getPlayerName() {
+		return "server";
+	}
+
 }