+++ 1.10.2/src/main/java/com/stateshifterlabs/achievementbooks/networking/ToggleAchievementMessage.java
@@ -8,8 +8,8 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 public class ToggleAchievementMessage implements IMessage {
 	private final ToggleAchievementMessageBase base = new ToggleAchievementMessageBase(new BufferUtilities());
 
-	public ToggleAchievementMessage withData(String bookName, int achievmenetId) {
-		base.withData(bookName, achievmenetId);
+	public ToggleAchievementMessage withData(String username, String bookName, int achievmenetId) {
+		base.withData(username, bookName, achievmenetId);
 		return this;
 	}
 
@@ -29,6 +29,10 @@ public class ToggleAchievementMessage implements IMessage {
 
 	public String bookName() {
 		return base.bookName();
+	}
+
+	public String playerName() {
+		return base.playerName();
 	}
 
 }