+++ 1.7.10/src/main/java/com/stateshifterlabs/achievementbooks/CommonProxy.java
@@ -6,4 +6,8 @@ public class CommonProxy {
 		return "";
 	}
 
+	public String getPlayerName() {
+		return "server";
+	}
+
 }