+++ 1.7.10/src/main/java/com/stateshifterlabs/achievementbooks/commands/MainCommand.java
@@ -52,11 +52,22 @@ public class MainCommand extends CommandBase {
 
 	@Override
 	public List addTabCompletionOptions(ICommandSender sender, String[] strings) {
+
 		List<String> tabOptions = new ArrayList<String>();
 
 		if (strings.length == 1) {
 			for (CommandBase command : subCommands) {
-				tabOptions.add(command.getCommandName());
+				if(command.getCommandName().substring(0, strings[0].length()).equalsIgnoreCase(strings[0])) {
+					tabOptions.add(command.getCommandName());
+				}
+			}
+		}
+
+		if (strings.length > 1) {
+			for (CommandBase command : subCommands) {
+				if(command.getCommandName().equalsIgnoreCase(strings[0])) {
+					return command.addTabCompletionOptions(sender, strings);
+				}
 			}
 		}