+++ core/src/main/java/com/stateshifterlabs/achievementbooks/commands/Bind.java
@@ -1,0 +1,8 @@
+package com.stateshifterlabs.achievementbooks.commands;
+
+public class Bind {
+
+	public static final int PERMISSION = 4;
+	public static final String NAME = "bind";
+	public static final String USAGE = "bind <player>";
+}