+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-The very first official release of Achievement Books.
+- Added /ab bind <player> for server owners to be able to maintain player progress
+- Tab completion now narrows options as you type more characters
 
 For more information, head to the [project page](https://github.com/meza/achievementbooks) on how to use it.
 Also, feedback, issues, pull requests are all appreciated. Let me know what you would want to see in this mod.